[backend] fix account deletion if user has no profile picture

+++ logigator-backend/src/services/user.service.ts
@@ -265,7 +265,9 @@ export class UserService {
 			}
 		});
 		await projRepo.remove(projects);
-		await profilePicRepo.remove(user.image);
+		if (user.image) {
+			await profilePicRepo.remove(user.image);
+		}
 	}
 
 }
